move DockingDataset class to docking.schema
